# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Clientes",
        "description": "Documentação interativa do nosso CRUD de clientes com integração de CEP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista todos os clientes",
                "operationId": "2b23ac696cfd81c23807c2c1fd63b0b0",
                "responses": {
                    "200": {
                        "description": "Lista de clientes retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastra um novo cliente",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "document",
                                    "cep"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "01001000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação ou CEP inválido"
                    }
                }
            }
        },
        "/api/customers/{customer}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Busca os detalhes de um cliente específico",
                "operationId": "dab8c06ba11e633bb424e2fca1f0fbac",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cliente retornados com sucesso"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza os dados de um cliente",
                "operationId": "f28e50b8a6a110ecab6f2aa45eca980b",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Editado"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "81280120"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado com sucesso"
                    },
                    "400": {
                        "description": "Nenhum dado válido fornecido"
                    },
                    "422": {
                        "description": "Erro de validação ou CEP inválido"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Remove um cliente do sistema",
                "operationId": "eae002fbdbb9a6ef1275605e346d807f",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "ID do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente deletado com sucesso (Sem conteúdo)"
                    },
                    "404": {
                        "description": "Cliente não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "Clientes"
        }
    ]
}